pyproject.toml
[project]
name = "videotranscript"
version = "0.1.0"
description = ""
authors = [
    {name = "Utopi-a",email = "bridgestone.amatsu0228@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "transformers (>=4.48.1,<5.0.0)",
    "accelerate (>=1.3.0,<2.0.0)",
    "whisper (>=1.1.10,<2.0.0)",
    "llvmlite (>=0.44.0,<0.45.0)",
    "numba (>=0.61.0,<0.62.0)",
    "openai-whisper @ git+https://github.com/openai/whisper.git",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "ffmpeg (>=1.4,<2.0)",
    "moviepy (==1.0.3)",
    "torch (>=2.5.1,<3.0.0)",
    "torchvision (>=0.20.1,<0.21.0)",
    "torchaudio (>=2.5.1,<3.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "numpy (==2.1.3)",
    "librosa (>=0.10.2.post1,<0.11.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "scipy (>=1.15.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "torch_cu124"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "torch_cu124"}
torchvision = {source = "torch_cu124"}
torchaudio = {source = "torch_cu124"}
